+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Next
 
+### Features
+* Added total cpu and memory usage to builder logs.
+
 ### Fixes
 * Fixed aftermarket notification bug for expired claim bids.
 * Fixed not being able to scroll when sale and bid information doesn't fit on low resolution